adding and removing resources (sync not working perfectly)

+++ kubectl/index.ts
@@ -1,5 +1,7 @@
+import * as fs from "fs-extra";
 import * as shell from "shelljs";
 import * as yaml from "js-yaml";
+import * as _ from "lodash";
 
 export const all = (namespace: string) => {
   const response = execShell(`kubectl get all -n ${namespace} -o yaml`);
@@ -12,7 +14,17 @@ export const apply = (path: string) => {
 };
 
 export const remove = (path: string) => {
-  return execShell(`kubectl delete -f ${path}`);
+  if (fs.existsSync(path)) {
+    return execShell(`kubectl delete -f ${path}`);
+  }
+
+  const parts = _.split(path, "/");
+  const name = _.replace(parts[parts.length - 1], ".yml", "");
+  const kind = _.split(parts[parts.length - 2], "_")[0];
+  const namespace = parts[parts.length - 3];
+
+  console.log(`kubectl delete ${kind} -n ${namespace} ${name}`);
+  return execShell(`kubectl delete ${kind} -n ${namespace} ${name}`);
 };
 
 const execShell = (script: string) => {
